requirements.txt
Django==1.7.5
defusedxml==0.4.1
django-allauth==0.18.0
django-braces==1.4.0
django-oauth-toolkit==0.7.2
djangorestframework==3.0.2
oauthlib==0.7.2
psycopg2==2.5.4
python3-openid==3.0.5
pytz==2014.10
requests==2.5.1
selenium==2.44.0
shortuuid==0.4.2
six==1.9.0
